requirements.txt
fastapi
uvicorn
python-multipart
requests
jinja2
httpx
openai
python-dotenv
